+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 
 ### Added
 
+- Token configuration.
+- JSON configuration support.
+- YAML configuration support.
 
 [unreleased]: https://github.com/slok/simple-ingress-external-auth/compare/v0.1.0...HEAD
 [v0.1.0]: https://github.com/slok/simple-ingress-external-auth/releases/tag/v0.1.0